requirements.txt
torch
transformers
pyttsx3
SpeechRecognition
pyaudio
numpy==1.26.4
# For GPU support, llama-cpp-python needs to be installed with specific flags.
# Please see the installation instructions.t
llama-cpp-python
# For advanced VAD and high-quality TTS
webrtcvad-wheels
# For screen capture
Pillow
# For finding the game window on Windows
pywin32
# For web search
google-api-python-client
# For advanced memory (vector database)
chromadb
sentence-transformers

pygame
twitchio
style-bert-vits2
soundfile
psutil
websocket-client
ollama
aiohttp
bs4
numba
g2p-en

#python -m pip install torch==2.8.0 torchvision==0.23.0 torchaudio==2.8.0 --index-url https://download.pytorch.org/whl/xpu
#python -m pip install intel-extension-for-pytorch==2.8.10+xpu --index-url https://pytorch-extension.intel.com/release-whl/stable/xpu/us/

intel_npu_acceleration_library